Combine annotation files
