+++ src/debrief/index.js
@@ -1,10 +1,7 @@
 // @flow strict
 
-import { annotate } from './annotate';
-import { indent } from './utils';
-import serialize from './serialize';
-import summarize from './summarize';
-import type { Annotation } from './Annotation';
-
-export type { Annotation };
-export { annotate, indent, serialize, summarize };
+export { annotate } from './Annotation';
+export { indent } from './utils';
+export { summarize } from './summarize';
+export { serialize } from './serialize';
+export type { Annotation } from './Annotation';
